requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
beautifulsoup4
tiktoken
chromadb
python-dotenv
langgraph
fastapi
uvicorn
sentence-transformers
